+++ src/app/store/models/reducers/tokenState/saveTheme.ts
@@ -33,7 +33,7 @@ export function saveTheme(state: TokenState, data: Payload): TokenState {
   });
 
   const newActiveTheme = state.activeTheme;
-  if (isActiveTheme) {
+  if (!isActiveTheme) {
     Object.keys(newActiveTheme).forEach((group) => {
       if (newActiveTheme[group] === themeId) {
         delete newActiveTheme[group];