+++ src/app/store/useTokens.test.tsx
@@ -19,6 +19,7 @@ import { UpdateMode } from '@/constants/UpdateMode';
 import { NodeInfo } from '@/types/NodeInfo';
 import { Properties } from '@/constants/Properties';
 import { INTERNAL_THEMES_NO_GROUP } from '@/constants/InternalTokenGroup';
+import { TokensContext } from '@/context';
 
 type GetFormattedTokensOptions = {
   includeAllTokens: boolean;
@@ -101,6 +102,16 @@ const resolvedTokens: AnyTokenList = [
       textDecoration: 'none',
       textCase: 'none',
     },
+    resolvedValueWithReferences: {
+      fontFamily: 'Inter',
+      fontWeight: 'Regular',
+      lineHeight: 'AUTO',
+      fontSize: '14',
+      letterSpacing: '0%',
+      paragraphSpacing: '0',
+      textDecoration: 'none',
+      textCase: 'none',
+    },
   },
   {
     name: 'typography.alias',
@@ -116,6 +127,16 @@ const resolvedTokens: AnyTokenList = [
       textCase: 'none',
     },
     rawValue: '{typography.headlines.small}',
+    resolvedValueWithReferences: {
+      fontFamily: 'Inter',
+      fontWeight: 'Regular',
+      lineHeight: 'AUTO',
+      fontSize: '14',
+      letterSpacing: '0%',
+      paragraphSpacing: '0',
+      textDecoration: 'none',
+      textCase: 'none',
+    },
   },
   {
     name: 'font-family.serif',
@@ -160,6 +181,14 @@ const resolvedTokens: AnyTokenList = [
       color: '#000000',
       type: BoxShadowTypes.DROP_SHADOW,
     },
+    resolvedValueWithReferences: {
+      x: '2',
+      y: '2',
+      blur: '2',
+      spread: '2',
+      color: '#000000',
+      type: BoxShadowTypes.DROP_SHADOW,
+    },
   },
   {
     name: 'typography.boxshadow.alias',
@@ -173,12 +202,19 @@ const resolvedTokens: AnyTokenList = [
       type: BoxShadowTypes.DROP_SHADOW,
     },
     rawValue: '{typography.headlines.boxshadow}',
+    resolvedValueWithReferences: {
+      x: 2,
+      y: 2,
+      blur: 2,
+      spread: 2,
+      color: '#000000',
+      type: BoxShadowTypes.DROP_SHADOW,
+    },
   },
   {
     name: 'font-weight.regular',
     type: TokenTypes.FONT_WEIGHTS,
     value: 'Regular',
-    default: '400',
     rawValue: 'Regular',
   },
   {
@@ -186,7 +222,6 @@ const resolvedTokens: AnyTokenList = [
     type: TokenTypes.FONT_WEIGHTS,
     value: 'Regular',
     rawValue: '{font-weight.regular}',
-    default: '400',
   },
   {
     name: 'font-style.normal',
@@ -342,13 +377,19 @@ describe('useToken test', () => {
   it('remapTokensInGroup', async () => {
     const messageSpy = jest.spyOn(AsyncMessageChannel.ReactInstance, 'message');
     mockConfirm.mockImplementation(() => Promise.resolve({ data: ['selection'], result: true }));
+    const mockTokensToRename = [
+      {
+        oldName: 'old.padding-start',
+        newName: 'new.padding-start',
+      },
+    ];
     await act(async () => {
-      await result.current.remapTokensInGroup({ oldGroupName: 'old.', newGroupName: 'new.' });
+      await result.current.remapTokensInGroup({ oldGroupName: 'old.', newGroupName: 'new.', tokensToRename: mockTokensToRename });
     });
     expect(messageSpy).toBeCalledWith({
       type: AsyncMessageTypes.BULK_REMAP_TOKENS,
-      oldName: 'old.',
-      newName: 'new.',
+      oldName: 'old.padding-start',
+      newName: 'new.padding-start',
       updateMode: UpdateMode.SELECTION,
     });
   });
@@ -423,6 +464,10 @@ describe('useToken test', () => {
               fontFamily: 'Inter',
               fontWeight: 'Bold',
             },
+            resolvedValueWithReferences: {
+              fontFamily: 'Inter',
+              fontWeight: 'Bold',
+            },
             type: 'typography',
             value: {
               fontFamily: 'Inter',
@@ -458,6 +503,10 @@ describe('useToken test', () => {
             fontFamily: 'Inter',
             fontWeight: 'Bold',
           },
+          resolvedValueWithReferences: {
+            fontFamily: 'Inter',
+            fontWeight: 'Bold',
+          },
           type: 'typography',
           value: {
             fontFamily: 'Inter',
@@ -470,13 +519,12 @@ describe('useToken test', () => {
     });
     it('rename styles from current theme', async () => {
       await act(async () => {
-        await result.current.renameStylesFromTokens({ oldName: 'old', newName: 'new', parent: 'global' });
+        await result.current.renameStylesFromTokens([{ oldName: 'old', newName: 'new' }], 'global');
       });
 
       expect(messageSpy).toBeCalledWith({
         type: AsyncMessageTypes.RENAME_STYLES,
-        oldName: 'old',
-        newName: 'new',
+        tokensToRename: [{ oldName: 'old', newName: 'new' }],
         parent: 'global',
         settings: store.getState().settings,
       });
@@ -527,6 +575,130 @@ describe('useToken test', () => {
           removeStyle: true,
         },
         settings: store.getState().settings,
+      });
+    });
+  });
+
+  describe('createVariables', () => {
+    const tokenMockStore = createMockStore({
+      tokenState: {
+        usedTokenSet: { global: TokenSetStatus.ENABLED, light: TokenSetStatus.ENABLED },
+        activeTheme: {
+          [INTERNAL_THEMES_NO_GROUP]: 'light',
+        },
+        themes: [{
+          id: 'light', name: 'Light', selectedTokenSets: {}, $figmaStyleReferences: {},
+        }],
+        tokens: {
+          global: [{ name: 'white', value: '#ffffff', type: TokenTypes.COLOR }, { name: 'headline', value: { fontFamily: 'Inter', fontWeight: 'Bold' }, type: TokenTypes.TYPOGRAPHY }, { name: 'shadow', value: '{shadows.default}', type: TokenTypes.BOX_SHADOW }],
+          light: [
+            { name: 'bg.default', value: '#ffffff', type: TokenTypes.COLOR },
+            { name: 'borderRadius-1', value: '{base.base-0} {base.base-25}', type: TokenTypes.BORDER_RADIUS },
+            { name: 'borderRadius-2', value: '{base.base-50} {base.base-125}', type: TokenTypes.BORDER_RADIUS },
+            { name: 'borderRadius-3', value: '{base.base-150}', type: TokenTypes.BORDER_RADIUS },
+            { name: 'spacing-1', value: '{base.base-50} {base.base-150}', type: TokenTypes.SPACING },
+            { name: 'spacing-2', value: '{base.base-125} {base.base-50}', type: TokenTypes.SPACING },
+            { name: 'spacing-3', value: '{base.base-50}', type: TokenTypes.SPACING },
+            { name: 'borderWidth-1', value: '{base.base-350} {base.base-187} {base.base-50}', type: TokenTypes.BORDER_WIDTH },
+            { name: 'borderWidth-2', value: '{base.base-50} {base.base-187} {base.base-150}', type: TokenTypes.BORDER_WIDTH },
+            { name: 'borderWidth-3', value: '{base.base-187}', type: TokenTypes.BORDER_WIDTH },
+            { name: 'text', value: 'text value', type: TokenTypes.TEXT },
+            { name: 'shadow.base', value: '#1450c8', type: TokenTypes.COLOR },
+            {
+              value: '{color.slate.50}',
+              type: 'color',
+              name: 'shadow.strong',
+              $extensions: {
+                'studio.tokens': {
+                  modify: {
+                    type: 'alpha',
+                    value: '.3',
+                    space: 'lch',
+                  },
+                },
+              },
+            },
+            {
+              value: '{color.slate.50}',
+              type: 'color',
+              name: 'shadow.intense',
+              $extensions: {
+                'studio.tokens': {
+                  modify: {
+                    type: 'alpha',
+                    value: '.4',
+                    space: 'lch',
+                  },
+                },
+              },
+            },
+          ],
+        },
+      },
+    });
+
+    const customTokens = {
+      resolvedTokens,
+    };
+
+    beforeEach(() => {
+      resetStore();
+      result = renderHook(() => useTokens(), {
+        wrapper: ({ children }: { children?: React.ReactNode }) => (
+          <Provider store={tokenMockStore}>
+            <TokensContext.Provider value={customTokens}>
+              {children}
+            </TokensContext.Provider>
+            ,
+          </Provider>
+        ),
+      }).result;
+    });
+
+    it('skip multi value tokens and respect values ', async () => {
+      const multiValueFilteredTokens = result.current.filterMultiValueTokens();
+
+      expect(multiValueFilteredTokens).toStrictEqual({
+        global: [
+          { name: 'white', value: '#ffffff', type: TokenTypes.COLOR },
+          { name: 'headline', value: { fontFamily: 'Inter', fontWeight: 'Bold' }, type: TokenTypes.TYPOGRAPHY },
+          { name: 'shadow', value: '{shadows.default}', type: TokenTypes.BOX_SHADOW }],
+        light: [
+          { name: 'bg.default', value: '#ffffff', type: TokenTypes.COLOR },
+          { name: 'borderRadius-3', value: '{base.base-150}', type: TokenTypes.BORDER_RADIUS },
+          { name: 'spacing-3', value: '{base.base-50}', type: TokenTypes.SPACING },
+          { name: 'borderWidth-3', value: '{base.base-187}', type: TokenTypes.BORDER_WIDTH },
+          { name: 'text', value: 'text value', type: TokenTypes.TEXT },
+          { name: 'shadow.base', value: '#1450c8', type: TokenTypes.COLOR },
+          {
+            value: '#f8fafc',
+            type: 'color',
+            name: 'shadow.strong',
+            $extensions: {
+              'studio.tokens': {
+                modify: {
+                  type: 'alpha',
+                  value: '.3',
+                  space: 'lch',
+                },
+              },
+            },
+          },
+          {
+            value: '#f8fafc',
+            type: 'color',
+            name: 'shadow.intense',
+            $extensions: {
+              'studio.tokens': {
+                modify: {
+                  type: 'alpha',
+                  value: '.4',
+                  space: 'lch',
+                },
+              },
+            },
+          },
+        ],
       });
     });
   });