+++ src/figmaStorage/InitialLoadProperty.ts
@@ -1,0 +1,6 @@
+import { FigmaStorageProperty, FigmaStorageType } from './FigmaStorageProperty';
+
+export const InitialLoadProperty = new FigmaStorageProperty<boolean>(
+  FigmaStorageType.CLIENT_STORAGE,
+  'initialLoad',
+);